Fixed the layout for the Code Editor
