+++ package.json
@@ -39,6 +39,7 @@
     "@types/react-dom": "19.0.2",
     "@vercel/analytics": "^1.3.1",
     "@vercel/speed-insights": "^1.0.12",
+    "add": "^2.0.6",
     "autoprefixer": "10.4.14",
     "class-variance-authority": "^0.7.0",
     "classnames": "^2.3.2",
@@ -59,14 +60,16 @@
     "re-resizable": "^6.9.9",
     "react": "19.0.0",
     "react-dom": "19.0.0",
-    "react-resizable-panels": "^0.0.54",
+    "react-resizable-panels": "^2.1.7",
     "rehype-highlight": "^7.0.0",
     "remark-gfm": "^4.0.0",
+    "sandpack-file-explorer": "^0.0.7",
     "tailwind-merge": "^2.5.2",
     "tailwindcss": "3.3.3",
     "tailwindcss-animate": "^1.0.7",
     "typescript": "5.1.6",
-    "use-debounce": "^10.0.3"
+    "use-debounce": "^10.0.3",
+    "yarn": "^1.22.22"
   },
   "devDependencies": {
     "@tailwindcss/typography": "^0.5.9",
